+++ src/app/page.tsx
@@ -3,7 +3,7 @@ import { redirect } from "next/navigation";
 
 import Recipes from "@/components/recipes/Recipes";
 
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/service/auth";
 
 export default async function HomePage() {
   const session = await getServerSession(authOptions);
